camera.py
from picamera.array import PiRGBArray
from picamera import PiCamera
import time
import cv2

def initialize_camera():
    camera = PiCamera()
    camera.resolution = (320, 240)
    camera.framerate = 32
    rawCapture = PiRGBArray(camera, size=(320, 240))
    time.sleep(0.1)  # Allow the camera to warm up
    return camera, rawCapture

def capture_image(camera, rawCapture):
    camera.capture(rawCapture, format="bgr")
    image = rawCapture.array
    rawCapture.truncate(0)
    return image

def process_image(image):
    # Process the image to recognize lanes
    # Placeholder for image processing code
    pass

# Example usage
camera, rawCapture = initialize_camera()
while True:
    image = capture_image(camera, rawCapture)
    process_image(image)