requirements.txt
opencv-python
RPi.GPIO
picamera
Flask